MGM: Remove quota for logical bytes as well if set
Old versions of EOS will set/remove only userbytes/groupbytes. New
versions will set also logical bytes. To allow old/new versions of
EOS to be compatible. We set by logical bytes and set the raw bytes
based on the space factor and logical bytes. When removing, to be
able to remove both old and new quota style quota nodes, we need to
do it by raw bytes, and if a logical quota also exists, delete that
as well.